p2p, refactor: pass and use uint16_t CService::port as uint16_t

+++ src/net.h
@@ -229,7 +229,7 @@ extern std::string strSubVersion;
 
 struct LocalServiceInfo {
     int nScore;
-    int nPort;
+    uint16_t nPort;
 };
 
 extern RecursiveMutex cs_mapLocalHost;
